Add option to display all eMail-Addresses for a contact
It is possible to store more than one eMail-Address for a contact in a
VCard. This adds a new option '-a'/'--all-addresses' that will cause
the script to print all addresses instead of only the first one.

In mutt this will cause a selection-list to be displayed if more than
one address is printed for a query.

Signed-off-by: Benjamin Block <bebl@mageta.org>