Make it possible to cache for more than one VCard-Directory
Because the script always only used one cache-file, and invalidated all
paths that where found in a previous run but not in the current run, it
was effectively only possible to cache the content of one VCard-Directory.

Improve this by simply appending a hash of the VCard-Directory path to the
cache-path, so that each directory will automatically get its unique
cache-file.

New paths for caches look like:
    ~/.cache/<XDIGIT*64>.vcs_query

Signed-off-by: Benjamin Block <bebl@mageta.org>