Prevent duplicates from being printed
Till now there was no mechanism that would filter out duplicates in the
read contact-data (a contact being the tuple of eMail-Address, Name,
and Description). Especially if multiple VCard-Directories are given
and/or the option '--all-addresses' is used, this can lead to unnecessary
verbose output.

Fixing this is not completely straight-forward, as duplicates can be
found in different VCards, while the same VCards also contain different
contacts form the duplicate. For example:

  VCard M:
    Contact A (eMail X, Name A, Description A)
    Contact B (eMail Y, Name A, Description A)
  VCard N:
    Contact A (eMail X, Name A, Description A)
    Contact C (eMail Z, Name A, Description A)

We can not simply dismiss VCard N completely, because it contains
Contact C, but we want to only print one version of Contact A, because
the other is a duplicate.

To fix this, we first collect all VCards from all given VCard-Directories.
While collecting these, we add all contained contacts into a set, which
will automatically discard all duplicates. Afterwards we create a sorted
list from the contact-set and do the filtering/printing on that list.

Further, the VCard class doesn't do the data-formatting anymore, but only
returns a named tuple. This way we don't have to handle "string-blobs" in
the main functions.

Signed-off-by: Benjamin Block <bebl@mageta.org>