Add option to sort the contacts according to the name first
The default is to sort the contacts first according to the eMail-Address,
then name, and last description. With the new option '-n'/'--sort-names'
the order is changed to name, eMail, and description. Especially if
'--all-addresses' is used this is better because it keeps the addresses
by the same person ordered together.

The sort when '--starting-matches' is used is not changed.

Signed-off-by: Benjamin Block <bebl@mageta.org>