Add version-information to VcardCache to detect outdated caches
If we somehow change the layout/content or the information we cache in
VcardCache, but don't delete the existing caches, we either get crashes
or outdated information.

First example being the use of the VCards NOTE type for Vcard.description.
If the new code is used with the old cache-file, the description-
information is not updated and stays empty, even if a NOTE type is defined
for that VCard.

To make maintenance a bit easier, we bundle the state that gets pickled
into the new _state property.

Also add more error-handling for the unpickle function _load(). This checks
now for OSError, pickle.UnpicklingError, and whether the cache is
outdated and/or invalid (this doesn't make it secure, just more robust!
You still must not use caches from untrusted sources).

Signed-off-by: Benjamin Block <bebl@mageta.org>