Make vCard parsing more robust and update documentation for it
Add error-handling to the parsing-calls for the found vCards. This includes
errors in encoding and errors encountered by the parser. As before, we try
to salvage what is salvageable, if we found complete vCards already, we
keep using them.

Update the documentation concerning vCards (the current RFC is 6350) and
some specialties of the vobject library that might not be obvious, and thus
confusing. Also add more references into the (new) RFC for further reading.

Add a "Limitations" section to the README.md for things that users should
know when using the script.

Signed-off-by: Benjamin Block <bebl@mageta.org>