Add second output-mode to support use in Vim completion-functions
Add a new option -m/--mode that lets the user select a mode of how the
output will be formatted. The default remains to be 'mutt'. A new mode
'vim' is added so the script can easily be used in vim completion-functions
to complete eMail-Addresses.

A example script for how to implement such a completion-function is added
to the README.md, along more documentation for this new output-mode.

Signed-off-by: Benjamin Block <bebl@mageta.org>